Use UnsafeAccessor
