+++ src/libraries/System.Runtime.Numerics/src/System.Runtime.Numerics.csproj
@@ -32,12 +32,10 @@
              Link="CoreLib\System\Collections\Generic\ValueListBuilder.cs" />
     <Compile Include="$(CommonPath)System\Text\ValueStringBuilder.cs"
              Link="CoreLib\System\Text\ValueStringBuilder.cs" />
-    <Compile Include="$(CommonPath)System\Number.Formatting.Common.cs"
-             Link="Common\System\Number.Formatting.Common.cs" />
+    <Compile Include="$(CommonPath)System\Number.Common.cs"
+             Link="Common\System\Number.Common.cs" />
     <Compile Include="$(CommonPath)System\Number.NumberBuffer.cs"
              Link="Common\System\Number.NumberBuffer.cs" />
-    <Compile Include="$(CommonPath)System\Number.Parsing.Common.cs"
-             Link="Common\System\Number.Parsing.Common.cs" />
   </ItemGroup>
 
   <ItemGroup>
