+++ src/libraries/System.Runtime.Numerics/src/System/Number.Polyfill.cs
@@ -1,8 +1,10 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
+// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using System.Collections.Generic;
 using System.Diagnostics;
 using System.Globalization;
+using System.Numerics;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
@@ -11,158 +13,82 @@ namespace System
     // Polyfill CoreLib internal interfaces and methods
     // Define necessary members only
 
-    internal interface IUtfChar<TSelf> :
-        IEquatable<TSelf>
-        where TSelf : unmanaged, IUtfChar<TSelf>
-    {
-        public static abstract TSelf CastFrom(byte value);
-
-        public static abstract TSelf CastFrom(char value);
-
-        public static abstract TSelf CastFrom(int value);
-
-        public static abstract TSelf CastFrom(uint value);
-
-        public static abstract TSelf CastFrom(ulong value);
-
-        public static abstract uint CastToUInt32(TSelf value);
-    }
-
-#pragma warning disable CA1067 // Polyfill only type
-    internal readonly struct Utf16Char(char ch) : IUtfChar<Utf16Char>
-#pragma warning restore CA1067
-    {
-        private readonly char value = ch;
-
-        public static Utf16Char CastFrom(byte value) => new((char)value);
-        public static Utf16Char CastFrom(char value) => new(value);
-        public static Utf16Char CastFrom(int value) => new((char)value);
-        public static Utf16Char CastFrom(uint value) => new((char)value);
-        public static Utf16Char CastFrom(ulong value) => new((char)value);
-        public static uint CastToUInt32(Utf16Char value) => value.value;
-        public bool Equals(Utf16Char other) => value == other.value;
-    }
-
-#pragma warning disable CA1067 // Polyfill only type
-    internal readonly struct Utf8Char(byte ch) : IUtfChar<Utf8Char>
-#pragma warning restore CA1067
-    {
-        private readonly byte value = ch;
-
-        public static Utf8Char CastFrom(byte value) => new(value);
-        public static Utf8Char CastFrom(char value) => new((byte)value);
-        public static Utf8Char CastFrom(int value) => new((byte)value);
-        public static Utf8Char CastFrom(uint value) => new((byte)value);
-        public static Utf8Char CastFrom(ulong value) => new((byte)value);
-        public static uint CastToUInt32(Utf8Char value) => value.value;
-        public bool Equals(Utf8Char other) => value == other.value;
-    }
-
     internal static partial class Number
     {
-        internal static bool AllowHyphenDuringParsing(this NumberFormatInfo info)
+        private const string SYSTEM_NUMBER_CORELIB = "System.Number, System.Private.CoreLib";
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        private static unsafe bool TryStringToNumber<TChar>(ReadOnlySpan<TChar> value, NumberStyles styles, ref NumberBuffer number, NumberFormatInfo info)
+            where TChar : unmanaged, IBinaryInteger<TChar>
         {
-            string negativeSign = info.NegativeSign;
-            return negativeSign.Length == 1 &&
-                   negativeSign[0] switch
-                   {
-                       '\u2012' or         // Figure Dash
-                       '\u207B' or         // Superscript Minus
-                       '\u208B' or         // Subscript Minus
-                       '\u2212' or         // Minus Sign
-                       '\u2796' or         // Heavy Minus Sign
-                       '\uFE63' or         // Small Hyphen-Minus
-                       '\uFF0D' => true,   // Fullwidth Hyphen-Minus
-                       _ => false
-                   };
+            Debug.Assert(typeof(TChar) == typeof(byte) || typeof(TChar) == typeof(char));
+            return typeof(TChar) == typeof(byte)
+                ? TryStringToNumberCore(null, Unsafe.BitCast<ReadOnlySpan<TChar>, ReadOnlySpan<byte>>(value), styles, Unsafe.AsPointer(ref number), info)
+                : TryStringToNumberCore(null, Unsafe.BitCast<ReadOnlySpan<TChar>, ReadOnlySpan<char>>(value), styles, Unsafe.AsPointer(ref number), info);
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static ReadOnlySpan<TChar> PositiveSignTChar<TChar>(this NumberFormatInfo info)
-            where TChar : unmanaged, IUtfChar<TChar>
+        [UnsafeAccessor(UnsafeAccessorKind.StaticMethod, Name = nameof(TryStringToNumberCore))]
+        private static extern unsafe bool TryStringToNumberCore([UnsafeAccessorType(SYSTEM_NUMBER_CORELIB)] object? obj, ReadOnlySpan<byte> value, NumberStyles styles, void* number, NumberFormatInfo info);
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        [UnsafeAccessor(UnsafeAccessorKind.StaticMethod, Name = nameof(TryStringToNumberCore))]
+        private static extern unsafe bool TryStringToNumberCore([UnsafeAccessorType(SYSTEM_NUMBER_CORELIB)] object? obj, ReadOnlySpan<char> value, NumberStyles styles, void* number, NumberFormatInfo info);
+
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        internal static unsafe TChar* UInt32ToDecChars<TChar>(TChar* bufferEnd, uint value, int digits)
+          where TChar : unmanaged, IBinaryInteger<TChar>
         {
-            Debug.Assert(typeof(TChar) == typeof(Utf16Char));
-            return MemoryMarshal.Cast<char, TChar>(info.PositiveSign);
+            Debug.Assert(typeof(TChar) == typeof(byte) || typeof(TChar) == typeof(char));
+            return typeof(TChar) == typeof(byte)
+                ? (TChar*)UInt32ToDecCharsCore(null, (byte*)bufferEnd, value, digits)
+                : (TChar*)UInt32ToDecCharsCore(null, (char*)bufferEnd, value, digits);
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static ReadOnlySpan<TChar> NegativeSignTChar<TChar>(this NumberFormatInfo info)
-            where TChar : unmanaged, IUtfChar<TChar>
+        [UnsafeAccessor(UnsafeAccessorKind.StaticMethod, Name = nameof(UInt32ToDecCharsCore))]
+        internal static extern unsafe byte* UInt32ToDecCharsCore([UnsafeAccessorType(SYSTEM_NUMBER_CORELIB)] object? obj, byte* bufferEnd, uint value, int digits);
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        [UnsafeAccessor(UnsafeAccessorKind.StaticMethod, Name = nameof(UInt32ToDecCharsCore))]
+        internal static extern unsafe char* UInt32ToDecCharsCore([UnsafeAccessorType(SYSTEM_NUMBER_CORELIB)] object? obj, char* bufferEnd, uint value, int digits);
+
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        internal static char ParseFormatSpecifier(ReadOnlySpan<char> format, out int digits)
+            => ParseFormatSpecifierCore(null, format, out digits);
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        [UnsafeAccessor(UnsafeAccessorKind.StaticMethod, Name = nameof(ParseFormatSpecifier))]
+        internal static extern char ParseFormatSpecifierCore([UnsafeAccessorType(SYSTEM_NUMBER_CORELIB)] object? obj, ReadOnlySpan<char> format, out int digits);
+
+
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        internal static unsafe string NumberToString(ref NumberBuffer number, char fmt, int digits, ReadOnlySpan<char> format, NumberFormatInfo info)
+            => NumberToStringCore(null, Unsafe.AsPointer(ref number), fmt, digits, format, info);
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        [UnsafeAccessor(UnsafeAccessorKind.StaticMethod, Name = nameof(NumberToStringCore))]
+        internal static extern unsafe string NumberToStringCore([UnsafeAccessorType(SYSTEM_NUMBER_CORELIB)] object? obj, void* number, char fmt, int digits, ReadOnlySpan<char> format, NumberFormatInfo info);
+
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        internal static unsafe bool FormatNumber<TChar>(ref NumberBuffer number, char fmt, int digits, ReadOnlySpan<char> format, NumberFormatInfo info, Span<TChar> destination, out int charsWritten)
         {
-            Debug.Assert(typeof(TChar) == typeof(Utf16Char));
-            return MemoryMarshal.Cast<char, TChar>(info.NegativeSign);
+            Debug.Assert(typeof(TChar) == typeof(byte) || typeof(TChar) == typeof(char));
+            return typeof(TChar) == typeof(byte)
+                ? FormatNumberCore(null, Unsafe.AsPointer(ref number), fmt, digits, format, info, Unsafe.BitCast<Span<TChar>, Span<byte>>(destination), out charsWritten)
+                : FormatNumberCore(null, Unsafe.AsPointer(ref number), fmt, digits, format, info, Unsafe.BitCast<Span<TChar>, Span<char>>(destination), out charsWritten);
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static ReadOnlySpan<TChar> CurrencySymbolTChar<TChar>(this NumberFormatInfo info)
-            where TChar : unmanaged, IUtfChar<TChar>
-        {
-            Debug.Assert(typeof(TChar) == typeof(Utf16Char));
-            return MemoryMarshal.Cast<char, TChar>(info.CurrencySymbol);
-        }
+        [UnsafeAccessor(UnsafeAccessorKind.StaticMethod, Name = nameof(FormatNumberCore))]
+        internal static extern unsafe bool FormatNumberCore([UnsafeAccessorType(SYSTEM_NUMBER_CORELIB)] object? obj, void* number, char fmt, int digits, ReadOnlySpan<char> format, NumberFormatInfo info, Span<byte> destination, out int charsWritten);
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static ReadOnlySpan<TChar> PercentSymbolTChar<TChar>(this NumberFormatInfo info)
-            where TChar : unmanaged, IUtfChar<TChar>
-        {
-            Debug.Assert(typeof(TChar) == typeof(Utf16Char));
-            return MemoryMarshal.Cast<char, TChar>(info.PercentSymbol);
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static ReadOnlySpan<TChar> PerMilleSymbolTChar<TChar>(this NumberFormatInfo info)
-            where TChar : unmanaged, IUtfChar<TChar>
-        {
-            Debug.Assert(typeof(TChar) == typeof(Utf16Char));
-            return MemoryMarshal.Cast<char, TChar>(info.PerMilleSymbol);
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static ReadOnlySpan<TChar> CurrencyDecimalSeparatorTChar<TChar>(this NumberFormatInfo info)
-            where TChar : unmanaged, IUtfChar<TChar>
-        {
-            Debug.Assert(typeof(TChar) == typeof(Utf16Char));
-            return MemoryMarshal.Cast<char, TChar>(info.CurrencyDecimalSeparator);
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static ReadOnlySpan<TChar> CurrencyGroupSeparatorTChar<TChar>(this NumberFormatInfo info)
-            where TChar : unmanaged, IUtfChar<TChar>
-        {
-            Debug.Assert(typeof(TChar) == typeof(Utf16Char));
-            return MemoryMarshal.Cast<char, TChar>(info.CurrencyGroupSeparator);
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static ReadOnlySpan<TChar> NumberDecimalSeparatorTChar<TChar>(this NumberFormatInfo info)
-            where TChar : unmanaged, IUtfChar<TChar>
-        {
-            Debug.Assert(typeof(TChar) == typeof(Utf16Char));
-            return MemoryMarshal.Cast<char, TChar>(info.NumberDecimalSeparator);
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static ReadOnlySpan<TChar> NumberGroupSeparatorTChar<TChar>(this NumberFormatInfo info)
-            where TChar : unmanaged, IUtfChar<TChar>
-        {
-            Debug.Assert(typeof(TChar) == typeof(Utf16Char));
-            return MemoryMarshal.Cast<char, TChar>(info.NumberGroupSeparator);
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static ReadOnlySpan<TChar> PercentDecimalSeparatorTChar<TChar>(this NumberFormatInfo info)
-            where TChar : unmanaged, IUtfChar<TChar>
-        {
-            Debug.Assert(typeof(TChar) == typeof(Utf16Char));
-            return MemoryMarshal.Cast<char, TChar>(info.PercentDecimalSeparator);
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static ReadOnlySpan<TChar> PercentGroupSeparatorTChar<TChar>(this NumberFormatInfo info)
-            where TChar : unmanaged, IUtfChar<TChar>
-        {
-            Debug.Assert(typeof(TChar) == typeof(Utf16Char));
-            return MemoryMarshal.Cast<char, TChar>(info.PercentGroupSeparator);
-        }
+        [UnsafeAccessor(UnsafeAccessorKind.StaticMethod, Name = nameof(FormatNumberCore))]
+        internal static extern unsafe bool FormatNumberCore([UnsafeAccessorType(SYSTEM_NUMBER_CORELIB)] object? obj, void* number, char fmt, int digits, ReadOnlySpan<char> format, NumberFormatInfo info, Span<char> destination, out int charsWritten);
     }
 }
